+++ fly-admin/src/main/java/fly/admin/repository/PostCommentRepository.java
@@ -3,6 +3,6 @@ package fly.admin.repository;
 import fly.admin.entity.model.PostComment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostCommentRepository extends JpaRepository<PostComment,Integer> {
-    int countByUserId(int userId);
+public interface PostCommentRepository extends JpaRepository<PostComment,Long> {
+    int countByUserId(Long userId);
 }
